+++ src/main.tsx
@@ -2,9 +2,12 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import App from "./components/App/App.tsx";
 import "./index.css";
+import { GameProvider } from "./context/GameContext.tsx";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
-    <App />
+    <GameProvider>
+      <App />
+    </GameProvider>
   </StrictMode>
 );
